[wip] auto install

+++ bin/evergreen.sh
@@ -5,6 +5,8 @@
 set -e
 set -vx
 
+mothership='mothership.alxb.us'
+
 cd /root
 
 opkg update
@@ -15,9 +17,19 @@ if [ ! -f '/root/resolv.conf' ]; then
   ln -f -s /root/resolv.conf /etc/resolv.conf
 fi
 
+ping -c2 $mothership
+[ $? -eq 0 ] || exit 1
+
 rm -rf evergreen
 git clone git://github.com/alexebird/evergreen.git
-cd evergreen
+[ $? -eq 0 ] || exit 2
 
-# should be in an init script
+cd evergreen
+local ifconfig_f='/root/ifconfig.txt'
+ifconfig | tr '\n' '$' > $ifconfig_f
+curl -XPOST $mothership:8889/arduino -d @$ifconfig_f
+[ $? -eq 0 ] || exit 3
+rm -f $ifconfig_f
+
+# should be in an init script maybe...
 ruby ./bin/evergreen.rb
